+++ platform/viewer/public/config/default.js
@@ -43,12 +43,12 @@ const defaultConfig = {
     // ~ Global
     {
       commandName: 'incrementActiveViewport',
-      label: 'Next Image Viewport',
+      label: 'Next Viewport',
       keys: ['right'],
     },
     {
       commandName: 'decrementActiveViewport',
-      label: 'Previous Image Viewport',
+      label: 'Previous Viewport',
       keys: ['left'],
     },
     // Supported Keys: https://craig.is/killing/mice